WIP: First draft packaging as a Nix flake
Using poetry2nix, and correspondingly for the moment switching from setuptools to poetry. Liable to change this going forward (perhaps keeping the main project's packaging as-is and having a separate project that treats lmql as a dependency).

Also in nonfunctional work-in-progress draft state is javascript packaging for the playground and live server.